test-support: backend-aware milestone harness (TestDb)
Introduces crates/test-support with a unified TestDb that abstracts
database backend (SQLite/PostgreSQL). Converts all 19 milestone test
files + revision_cache to use the unified Harness with tdb: TestDb.

- SQLite: scratch DB via TestDb::connect_with_dir (unchanged behavior)
- PostgreSQL: admin DB + template DB via LOREHAVEN_TEST_PG_URL env var

Fixes the 'SQLite-forgiving, PG-strict' silent-defect class by making
the same test code run against both backends.

+++ migrations/postgres/0012_discovery.sql
@@ -17,7 +17,7 @@ CREATE TABLE recipes (
     owner TEXT NOT NULL,
     name TEXT NOT NULL,
     document TEXT NOT NULL,
-    is_public INTEGER NOT NULL DEFAULT 0,
+    is_public BIGINT NOT NULL DEFAULT 0,
     created_at TEXT NOT NULL
 );
